+++ lib/src/clixon_yang_module.c
@@ -94,7 +94,7 @@ yang_modules_init(clicon_handle h)
 	goto done;
     }
     /* Ensure revision exists is set */
-    if (yang_spec_parse_module(h, "ietf-yang-library", CLIXON_DATADIR, NULL, yspec)< 0)
+    if (yang_spec_parse_module(h, "ietf-yang-library", CLIXON_DATADIR, NULL, yspec, NULL)< 0)
 	goto done;
     /* Find revision */
     if (yang_modules_revision(h) == NULL){